Decoupled ComputeForm component from shared functionality
