+++ src/steps/05-PerformanceRequirements/DOW/ComputeForm.vue
@@ -1,5 +1,5 @@
 <template>
-  <v-form ref="computeForm">
+  <div>
     <h1 class="page-header mb-3" tabindex="-1">
       <span v-if="firstTimeHere">Let’s start by gathering your Compute requirements</span>
       <span v-else>
@@ -147,54 +147,6 @@
       :availablePeriodCheckboxItems="availablePeriodCheckboxItems"
       index="1"
     />
-    <!-- <ATATRadioGroup
-      class="copy-max-width mb-10 mt-4"
-      ref="NeededForEntireDuration"
-      id="NeededForEntireDuration"
-      legend="Is this instance needed for the entire duration of your task order?"
-      :items="requirementOptions"
-      :value.sync="_computeData.entireDuration"
-      :rules="[
-        $validators.required('Please select an option to specify your requirement’s duration.')
-      ]"
-    />
-    <div v-if="_computeData.entireDuration === 'NO'">
-      <p id="PeriodsLabel" class="_checkbox-group-label">
-        In which base and/or option periods do you need this requirement?
-      </p>
-      <ATATCheckboxGroup
-        id="PeriodsCheckboxes"
-        ref="periodsCheckboxes"
-        aria-describedby="PeriodsLabel"
-        :value.sync="_computeData.periodsNeeded"
-        :items="availablePeriodCheckboxItems"
-        :card="false"
-        :disabled="isPeriodsDataMissing"
-        :rules="[
-          $validators.required('Please select at least one base or option period' +
-            ' to specify your requirement’s duration level.')
-        ]"
-        class="copy-max-width"
-      />
-      <ATATAlert
-        id="PeriodRequirementsAlert"
-        v-show="isPeriodsDataMissing === true"
-        type="warning"
-        class="copy-max-width mb-10"
-      >
-        <template v-slot:content>
-          <p class="mb-0" id="PeriodIntro">
-            Your period of performance details are missing. To select specific base or
-            option periods for this requirement,
-            <router-link
-              id="ContractDetailsLink"
-              :to="{name: routeNames.PeriodOfPerformance}"
-            >revisit the Contract Details section
-            </router-link>
-          </p>
-        </template>
-      </ATATAlert>
-    </div> -->
 
     <hr />
 
@@ -304,23 +256,13 @@
       </v-col>
     </v-row>
 
-    <ClassificationsModal 
-      :showDialog="showDialog"
-      @cancelClicked="modalCancelClicked"
-      @okClicked="classificationLevelsChanged"
-      :modalSelectedOptions.sync="modalSelectedOptions"
-      :modalSelectionsOnOpen="modalSelectionsOnOpen"
-      :modalCheckboxItems="modalCheckboxItems"
-      :IL6SysId="IL6SysId"
-      :isIL6Selected.sync="isIL6Selected"
-    />
-
-  </v-form>
+  </div>
 </template>
 
 <script lang="ts">
 import Vue from "vue";
 import { Component, Prop, PropSync, Watch } from "vue-property-decorator";
+import { ClassificationLevelDTO } from "@/api/models";
 
 import ATATAlert from "@/components/ATATAlert.vue";
 import ATATCheckboxGroup from "@/components/ATATCheckboxGroup.vue";
@@ -331,32 +273,15 @@ import ATATTextField from "@/components/ATATTextField.vue";
 import ATATTooltip from "@/components/ATATTooltip.vue"
 
 import DOWSubtleAlert from "./DOWSubtleAlert.vue";
-import ClassificationsModal from "./ClassificationsModal.vue";
 import EntireDuration from "./EntireDuration.vue"
-
-// import { routeNames } from "../../../router/stepper"
-import Periods from "@/store/periods";
-import { PeriodDTO } from "@/api/models";
-import classificationRequirements from "@/store/classificationRequirements";
-import Toast from "@/store/toast";
 
 import { 
   Checkbox, 
-  ComputeData,
+  OtherServiceOfferingData,
   RadioButton,
   SelectData,
-  ToastObj,
 } from "../../../../types/Global";
 
-import ClassificationRequirements from "@/store/classificationRequirements";
-import { ClassificationLevelDTO } from "@/api/models";
-
-import { 
-  buildClassificationCheckboxList, 
-  buildClassificationLabel,
-  createPeriodCheckboxItems,
-} from "@/helpers";
-import DescriptionOfWork from "@/store/descriptionOfWork";
 
 @Component({
   components: {
@@ -367,49 +292,31 @@ import DescriptionOfWork from "@/store/descriptionOfWork";
     ATATTextArea,
     ATATTextField,
     ATATTooltip,
-    ClassificationsModal,
     DOWSubtleAlert,
     EntireDuration,
-
   }
 })
 
 export default class ComputeForm extends Vue {
-  $refs!: {
-    computeForm: Vue & {
-      resetValidation: () => void;
-      errorBucket: string[];
-      reset: () => void;
-      validate: () => boolean;
-      errorBag: Record<number, boolean>;
-    },
-  };
-
-  @PropSync("computeData") public _computeData!: ComputeData;
-  @Prop() public isPeriodsDataMissing!: boolean;
+  @PropSync("computeData") public _computeData!: OtherServiceOfferingData;
+  @Prop() public firstTimeHere!: boolean;
   @Prop() public isClassificationDataMissing!: boolean;
+  @Prop() public isPeriodsDataMissing!: boolean;
+  @Prop() public avlClassificationLevelObjects!: ClassificationLevelDTO[];
+  @Prop() public singleClassificationLevelName!: string | undefined;
+  @Prop() public otherRegionValue!: string;
+  @Prop() public otherPerformanceTierValue!: string;
+  @Prop() public formHasErrors!: boolean;
+  @Prop() public formHasBeenTouched!: boolean;
+  @Prop() public classificationRadioOptions!: RadioButton[];
+  @Prop() public availablePeriodCheckboxItems!: Checkbox[];
+  @Prop() public validateOtherTierNow!: boolean;
+  @Prop() public validateOtherTierOnBlur!: boolean;
+  @Prop() public clearOtherTierValidation!: boolean;
 
-  public firstTimeHere = false;
-  // public showSubtleAlert = false;
-  // public routeNames = routeNames;
-  public modalSelectionsOnOpen: string[] = [];
-  public showDialog = false;
-  public modalSelectedOptions: string[] = [];
-  public modalCheckboxItems: Checkbox[] = [];
-  public isIL6Selected = false;
-  public IL6SysId = "";
-  public allClassificationLevels:ClassificationLevelDTO[] = [];
-  public avlClassificationLevelObjects: ClassificationLevelDTO[] = [];
-  public classificationRadioOptions: RadioButton[] = [];
-  public singleClassificationLevelName: string | undefined = "";
-
-  public classificationLevelToast: ToastObj = {
-    type: "success",
-    message: "Classification requirements updated",
-    isOpen: true,
-    hasUndo: false,
-    hasIcon: true,
-  };
+  public openModal(): void {
+    this.$emit("openModal");
+  }
 
   public EnvironmentTypeOptions: RadioButton[] = [
     {
@@ -457,37 +364,7 @@ export default class ComputeForm extends Vue {
     },
   ];
 
-  public otherRegionValue = "OtherRegion";
   public otherRegionValueRequiredMessage = "Please enter your other region(s).";
-
-  // public requirementOptions: RadioButton[] = [
-  //   {
-  //     id: "Yes",
-  //     label: "Yes",
-  //     value: "YES",
-  //   },
-  //   {
-  //     id: "No",
-  //     label: "No",
-  //     value: "NO",
-  //   },
-  // ];
-
-  // when user selects "YES", remove periods from needed array. 
-  // when user selects "NO", pre-select base period
-  @Watch("_computeData.entireDuration")
-  public entireDurationChanged(newVal: string): void {
-    debugger;
-    this._computeData.periodsNeeded = newVal === "NO"
-      ? [this.availablePeriodCheckboxItems[0].value]
-      : [];
-  }
-
-  public availablePeriodCheckboxItems: Checkbox[] = [];
-  public periodsDisabled = true;
-
-  // public isPeriodsDataMissing = false;
-  // public isClassificationDataMissing = false;
 
   public storageTypes: SelectData[] = [
     { text: "General Purpose SSD", value: "General Purpose SSD" },
@@ -496,10 +373,6 @@ export default class ComputeForm extends Vue {
     { text: "Offline", value: "Offline" },
     { text: "Other", value: "Other" },
   ];
-
-  public otherPerformanceTierValue = "OtherPerformance";
-  public otherPerformanceTierValueRequiredMessage 
-    = "Please enter your other performance tier.";
 
   public performanceTiers: RadioButton[] = [
     {
@@ -523,248 +396,8 @@ export default class ComputeForm extends Vue {
       value: "OtherPerformance",
     },
   ];
-
-  public clearOtherTierValidation = false;
-
-  public openModal(): void {
-    this.modalSelectionsOnOpen = this.modalSelectedOptions;
-    this.showDialog = true;
-  }
-
-  public modalCancelClicked(): void {
-    this.showDialog = false;
-  }
-
-  public setAvlClassificationLevels(): void {
-    this.classificationRadioOptions 
-      = this.createCheckboxOrRadioItems(this.avlClassificationLevelObjects, "Radio");
-  }
-
-  public checkSingleClassification(): void {
-    // if only one classification level selected in Contract Details or the 
-    // classifications modal, set it as the "selected" classification level
-    if (
-      this.avlClassificationLevelObjects.length === 1
-      && this.avlClassificationLevelObjects[0].sys_id
-    ) {
-      const classificationObj = this.avlClassificationLevelObjects[0];
-      this._computeData.classificationLevel = classificationObj.sys_id;
-      this.singleClassificationLevelName 
-        = buildClassificationLabel(classificationObj, "short");
-    }
-  }
-
-  public async classificationLevelsChanged(): Promise<void> {
-    this.showDialog = false;
-    this.avlClassificationLevelObjects = [];
-    this.modalSelectedOptions.forEach((sysId) => {
-      const classififcationObj = this.allClassificationLevels.find(obj => obj.sys_id === sysId);
-      if (classififcationObj) {
-        this.avlClassificationLevelObjects.push(classififcationObj);
-      }
-    });
-    this.setAvlClassificationLevels();
-
-    if (this.avlClassificationLevelObjects.length === 1) {
-      this.checkSingleClassification();
-    } else if (this._computeData.classificationLevel) {
-      // if the classification level that was selected was removed via the modal,
-      // clear out this._computeData.classificationLevel
-      const selectedSysId = this._computeData.classificationLevel;
-      if (this.modalSelectedOptions.indexOf(selectedSysId) === -1) {
-        this._computeData.classificationLevel = "";
-      }
-    }
-
-    await ClassificationRequirements.setSelectedClassificationLevels(
-      this.avlClassificationLevelObjects
-    );
-
-    Toast.setToast(this.classificationLevelToast);
-  }
-
-  // private createPeriodCheckboxItems(periods: PeriodDTO[]) {
-  //   // ensure sort order is correct
-  //   periods.sort((a, b) => a.option_order > b.option_order ? 1 : -1);
-    
-  //   const arr: Checkbox[] = [];
-  //   periods.forEach((period, i) => {
-  //     const label = i === 0 ? "Base period" : `Option period ${i}`;
-  //     const id = i === 0 ? "BASE" : `OPTION${i}`;
-  //     const option: Checkbox = {
-  //       id,
-  //       label,
-  //       value: period.sys_id || "",
-  //     };
-  //     arr.push(option);
-  //   })
-  //   return arr;
-  // }
-
-  private createCheckboxOrRadioItems(data: ClassificationLevelDTO[], idSuffix: string) {
-    idSuffix = idSuffix || "";
-    return data.length > 1 ? buildClassificationCheckboxList(data, idSuffix, false, false) : [];
-  }
-
-  public async setAvailableClassificationLevels(): Promise<void> {
-    this.avlClassificationLevelObjects 
-      = await ClassificationRequirements.getSelectedClassificationLevels();
-  }
-
-  public async loadOnEnter(): Promise<void> {
-    const computeObj = DescriptionOfWork.computeObject;
-    this.firstTimeHere = !computeObj.computeData || computeObj.computeData.length === 0;
-
-    // get classification levels selected in step 4 Contract Details
-    this.avlClassificationLevelObjects 
-      = await ClassificationRequirements.getSelectedClassificationLevels();
-    // set checked items in modal to classification levels selected in step 4 Contract Details
-    if (this.avlClassificationLevelObjects) {
-      this.avlClassificationLevelObjects.forEach((val) => {
-        this.modalSelectedOptions.push(val.sys_id || "")
-      });
-      this.checkSingleClassification();
-    }
-
-    // set available classification levels for radio buttons if > 1 level selected
-    await this.setAvailableClassificationLevels();
-    // get list of all possible classification levels to generate checkbox list and labels
-    this.allClassificationLevels
-      = await ClassificationRequirements.getAllClassificationLevels();
-    this.modalCheckboxItems 
-      = this.createCheckboxOrRadioItems(this.allClassificationLevels, "Modal");
-    const IL6Checkbox 
-      = this.modalCheckboxItems.find(e => e.label.indexOf("IL6") > -1);
-    this.IL6SysId = IL6Checkbox?.value || "";
-    
-    this.setAvlClassificationLevels();
-    this.checkSingleClassification();
-
-    this.availablePeriodCheckboxItems = await createPeriodCheckboxItems();
-    // if (periods && periods.length > 0) {
-    //   this.availablePeriodCheckboxItems = createPeriodCheckboxItems(periods);
-    // } else {
-    //   this.availablePeriodCheckboxItems = [
-    //     {
-    //       id: "BaseDisabled",
-    //       label: "Base period",
-    //       value: "",
-    //     }
-    //   ];
-    //   this.isPeriodsDataMissing = true;
-    // }
-    
-    // const classifications = await classificationRequirements.getSelectedClassificationLevels();
-    // this.isClassificationDataMissing = classifications.length === 0 ? true : false;
-
-    // this.showSubtleAlert 
-    //   = this.isPeriodsDataMissing || this.isClassificationDataMissing ? true : false;
-  }
-  public formHasBeenTouched = false;
-  public formHasErrors = false;
-  public errorBagValues: boolean[] = []
-  public hasErrorsOnLoad = false;
-  public validateOtherTierNow = false;
-  public validateOtherTierOnBlur = false;
-
-  public async mounted(): Promise<void> {
-    await this.loadOnEnter();
-    this.formHasBeenTouched = DescriptionOfWork.computeInstancesTouched.indexOf(
-      this._computeData.instanceNumber) > -1;
-    if (this.formHasBeenTouched) {
-      // user is editing an existing instance, validate on load
-      await this.validate();
-      this.$nextTick(async () => {
-        this.setErrorMessages();
-      });
-    } else {
-      this.validateOtherTierOnBlur = true;
-    }
-  };
-
-  public updated(): void {
-    const eb = this.$refs.computeForm.errorBag;
-    this.errorBagValues = Object.values(eb);
-  }
-
-  @Watch("errorBagValues")
-  public errorBagChange(): void {
-    this.$nextTick(() => {
-      const errorBag = Object.values(this.$refs.computeForm.errorBag);
-      this.formHasErrors = errorBag.includes(true);
-    });
-  }
-  
-  get Form(): Vue & { validate: () => boolean } {
-    return this.$refs.computeForm as Vue & { validate: () => boolean };
-  }
-
-  public async validate(): Promise<void> {
-    this.$nextTick(() => {
-      this.Form.validate();
-    });
-  }
-
-  private setErrorMessages(): void {
-    this.errorBagValues = Object.values(this.$refs.computeForm.errorBag);
-
-    const formChildren = this.$refs.computeForm.$children;
-    const inputRefs = [
-      "radioButtonGroup", "atatTextField", "atatTextArea", "atatSelect", "checkboxGroup",
-    ];
-
-    formChildren.forEach((child: any) => {
-      const refs = child.$refs;
-      const keys = Object.keys(refs);
-      keys.forEach((key: string) => {
-
-        if (inputRefs.indexOf(key) > -1) {
-          const childRef: any = child.$refs[key];
-          if (childRef[0]) {
-            if (childRef[0].attrs$["data-group-id"] === "PeriodsCheckboxes_Group"
-              && this._computeData.entireDuration.toLowerCase() === "no"
-              && this._computeData.periodsNeeded.length === 0
-            ) {
-              child.errorMessages.push(`Please select at least one base or option 
-                period to specify your requirement’s duration level.`);
-            }
-
-            if (childRef[0].attrs$["data-group-id"] === "Regions_Group"
-              && this._computeData.deployedRegions.indexOf(this.otherRegionValue) > -1
-              && this._computeData.deployedRegionsOther === ""
-            ) {
-              child.$refs["atatTextInput"][0].errorMessages.push(
-                'Please enter your other region(s).'
-              );
-            }
-          }
-
-          if (key === "radioButtonGroup" 
-            && child.$el.attributes.id.value.indexOf("PerformanceTier")
-            && this._computeData.performanceTier === this.otherPerformanceTierValue
-          ) {
-            if (this._computeData.performanceTierOther === "") {
-              this.validateOtherTierOnBlur = true;
-              this.validateOtherTierNow = true;
-            } else {
-              this.validateOtherTierOnBlur = false;
-              this.clearOtherTierValidation = true;
-            }
-          }
-
-          if (childRef && Object.prototype.hasOwnProperty.call(childRef, "errorBucket")) {
-            const errorBucket: string[] = childRef.errorBucket;
-            if (errorBucket.length) {
-              this.hasErrorsOnLoad = true;
-              errorBucket.forEach((error) => {
-                child.errorMessages.push(error);
-              });
-            }
-          }
-        }
-      });
-    });
-  }
+  public otherPerformanceTierValueRequiredMessage 
+    = "Please enter your other performance tier.";
 
   public regionTooltipText = `This is the geographic location where your public cloud 
     resources are located, e.g., within the continental U.S. (CONUS) or outside of the
