+++ src/steps/05-PerformanceRequirements/DOW/ServiceOfferingDetails.vue
@@ -376,11 +376,19 @@ export default class ServiceOfferingDetails extends Mixins(SaveOnLeave) {
   }
 
   private hasChanged(): boolean {
+    debugger;
     return hasChanges(this.instancesFormData, this.savedData);
   }
 
   protected async saveOnLeave(): Promise<boolean> {
     try {
+      debugger;
+      this.instancesFormData.forEach((instance, index) => {
+        if (instance.entireDuration.toLowerCase() === "yes") {
+          this.instancesFormData[index].selectedPeriods = [];
+        }
+      });
+
       if (this.hasChanged()) {
         // save to store
         await DescriptionOfWork.setOfferingDetails(this.instancesFormData);
