+++ src/steps/05-PerformanceRequirements/DOW/ServiceOfferings.vue
@@ -42,16 +42,14 @@
           </div>
         </v-col>
         
-        <v-col v-else-if="isCompute">
-          <ComputeForm 
-            :computeData.sync="computeData" 
+        <v-col v-else-if="isCompute || isGeneral">
+          <OtherOfferings 
+            :isCompute="isCompute"
+            :isGeneral="isGeneral"
+            :serviceOfferingData.sync="otherOfferingData" 
             :isPeriodsDataMissing="isPeriodsDataMissing"
             :isClassificationDataMissing="isClassificationDataMissing"
           />
-        </v-col>
-
-        <v-col v-else-if="isGeneral">
-          <GeneralXaaS :generalXaaSData.sync="generalXaaSData" />
         </v-col>
 
       </v-row>
@@ -64,7 +62,7 @@ import SaveOnLeave from "@/mixins/saveOnLeave";
 import { Component, Mixins, Watch } from "vue-property-decorator";
 
 import ATATCheckboxGroup from "@/components/ATATCheckboxGroup.vue";
-import ComputeForm from "./ComputeForm.vue";
+import OtherOfferings from "./OtherOfferings.vue";
 import DescriptionOfWork from "@/store/descriptionOfWork";
 import GeneralXaaS from "./GeneralXaaS.vue";
 import Periods from "@/store/periods";
@@ -74,18 +72,17 @@ import DOWSubtleAlert from "./DOWSubtleAlert.vue";
 
 import { 
   Checkbox, 
-  ComputeData, 
+  OtherServiceOfferingData, 
   DOWServiceOffering, 
-  GeneralXaaSData 
 } from "../../../../types/Global";
 import { getIdText } from "@/helpers";
 
 @Component({
   components: {
     ATATCheckboxGroup,
-    ComputeForm,
     DOWSubtleAlert,
     GeneralXaaS,
+    OtherOfferings,
   }
 })
 
@@ -116,7 +113,7 @@ export default class ServiceOfferings extends Mixins(SaveOnLeave) {
   public isGeneral = false;
   public isServiceOfferingList = true;
 
-  public computeData: ComputeData = {
+  public otherOfferingData: OtherServiceOfferingData = {
     instanceNumber: 1,
     environmentType: "",
     classificationLevel: "",
@@ -133,16 +130,9 @@ export default class ServiceOfferings extends Mixins(SaveOnLeave) {
     performanceTier: "",
     performanceTierOther: "",
     numberOfInstancesNeeded: "1",
-  }
-
-  public generalXaaSData: GeneralXaaSData = {
-    classificationLevel: "",
-    instanceNumber: 1,
     requirementTitle: "",
     descriptionOfNeed: "",
-    entireDuration: "",
-    periodsNeeded: [],
-  };
+  }
 
   public showSubtleAlert = false;
   public isPeriodsDataMissing = false;
@@ -200,15 +190,15 @@ export default class ServiceOfferings extends Mixins(SaveOnLeave) {
             obj => obj.instanceNumber === currentComputeInstanceNumber
           );
           if (computeData) {
-            this.computeData = computeData;
+            this.otherOfferingData = computeData;
           } else {
             const newComputeData 
               = await DescriptionOfWork.getComputeInstance(0);
             newComputeData.instanceNumber = currentComputeInstanceNumber;
-            this.computeData = newComputeData;
+            this.otherOfferingData = newComputeData;
           }
         } else {
-          this.computeData.instanceNumber = 1;
+          this.otherOfferingData.instanceNumber = 1;
           DescriptionOfWork.setCurrentComputeInstanceNumber(1);
         }
       }
@@ -235,7 +225,7 @@ export default class ServiceOfferings extends Mixins(SaveOnLeave) {
               { selectedOfferingSysIds: this.selectedOptions, otherValue: this.otherValueEntered }
             );
           } else if (this.isCompute) {
-            await DescriptionOfWork.setComputeData(this.computeData);
+            await DescriptionOfWork.setComputeData(this.otherOfferingData);
           }
         }
